# app/openapi_formatted.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetaMaster",
    "description": "Media Management Web Tool - Organize and manage your media library",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tv-shows": {
      "get": {
        "tags": [
          "TV Shows"
        ],
        "summary": "List Tv Shows",
        "description": "List TV shows with advanced filtering and sorting.\n\n- **genre**: Filter by genre (case-insensitive, optional)\n- **min_rating**: Minimum rating 0-10 (optional)\n- **max_rating**: Maximum rating 0-10 (optional)\n- **sort_by**: Sort field - title, rating, date_added (default: title)\n- **limit**: Number of items per page (1-100, default: 10)\n- **skip**: Offset from start (default: 0)",
        "operationId": "list_tv_shows_api_v1_tv_shows_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by genre (case-insensitive)",
              "title": "Genre"
            },
            "description": "Filter by genre (case-insensitive)"
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 0.0,
              "description": "Minimum rating (0-10)",
              "title": "Min Rating"
            },
            "description": "Minimum rating (0-10)"
          },
          {
            "name": "max_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 0.0,
              "description": "Maximum rating (0-10)",
              "title": "Max Rating"
            },
            "description": "Maximum rating (0-10)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort by: title, rating, date_added",
              "default": "title",
              "title": "Sort By"
            },
            "description": "Sort by: title, rating, date_added"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTVShowResponseWithFilters"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Create Tv Show",
        "description": "Create a new TV show",
        "operationId": "create_tv_show_api_v1_tv_shows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TVShowCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TVShowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-shows/{show_id}": {
      "get": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Get Tv Show",
        "description": "Get a specific TV show by ID",
        "operationId": "get_tv_show_api_v1_tv_shows__show_id__get",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TVShowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Update Tv Show",
        "description": "Update an existing TV show",
        "operationId": "update_tv_show_api_v1_tv_shows__show_id__put",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TVShowUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TVShowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Delete Tv Show",
        "description": "Delete a TV show",
        "operationId": "delete_tv_show_api_v1_tv_shows__show_id__delete",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-shows/{show_id}/seasons": {
      "get": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Get Tv Show Seasons",
        "description": "List all seasons for a TV show.\n\n- **show_id**: TV show ID\n- **limit**: Number of items per page (1-100, default: 10)\n- **offset**: Offset from start (default: 0)",
        "operationId": "get_tv_show_seasons_api_v1_tv_shows__show_id__seasons_get",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSeasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-shows/{show_id}/seasons/{season_id}/episodes": {
      "get": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Get Season Episodes",
        "description": "List all episodes for a season.\n\n- **show_id**: TV show ID\n- **season_id**: Season ID\n- **limit**: Number of items per page (1-100, default: 10)\n- **offset**: Offset from start (default: 0)",
        "operationId": "get_season_episodes_api_v1_tv_shows__show_id__seasons__season_id__episodes_get",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          },
          {
            "name": "season_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEpisodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tv-shows/{show_id}/sync-metadata": {
      "post": {
        "tags": [
          "TV Shows"
        ],
        "summary": "Sync Tv Show Metadata",
        "description": "Fetch TV show metadata from TVDB and update the TV show record.\n\nThis endpoint:\n1. Retrieves the TV show from the database\n2. Fetches updated metadata from TVDB using the show's TVDB ID\n3. Updates the TV show record with new metadata\n4. Returns the updated TV show information\n\nParameters:\n- **show_id**: ID of the TV show to sync\n\nReturns:\n- **success**: Whether sync was successful\n- **message**: Operation message\n- **show_id**: ID of the synced TV show\n- **updated_fields**: List of fields that were updated\n- **metadata**: Updated metadata",
        "operationId": "sync_tv_show_metadata_api_v1_tv_shows__show_id__sync_metadata_post",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Show Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedTVShowResponseWithFilters": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TVShowResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of TV shows"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of TV shows matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Items per page"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Offset from start"
          },
          "filters": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilterMetadata"
              }
            ],
            "description": "Applied filters and metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "filters"
        ],
        "title": "PaginatedTVShowResponseWithFilters",
        "description": "Paginated response for TV shows with filter metadata",
        "example": {
          "filters": {
            "applied_filters": {
              "genre": "Crime",
              "min_rating": 8.0
            },
            "sort_by": "rating",
            "total_results": 25
          },
          "items": [],
          "limit": 10,
          "offset": 0,
          "total": 25
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}